fix: PR #98 리뷰 반영 — QueryDSL 묵시적 JOIN 제거 및 코드 개선
- Expense: family_uuid 읽기 전용 필드 추가(insertable/updatable=false)로 QueryDSL 서브쿼리 묵시적 JOIN 방지
- FamilyRepositoryImpl: exp.family.uuid → exp.familyUuid, new QCategory("cat") → QCategory.category
- FamilyResponse: (int) 강제 캐스팅 → Math.toIntExact()로 오버플로우 방지
- FamilyWithCountsProjection: @NoArgsConstructor → @NoArgsConstructor(access = PRIVATE)

+++ src/main/java/com/bifos/accountbook/application/dto/family/FamilyResponse.java
@@ -70,9 +70,9 @@ public class FamilyResponse {
                          .monthlyBudget(projection.getMonthlyBudget())
                          .createdAt(projection.getCreatedAt())
                          .updatedAt(projection.getUpdatedAt())
-                         .memberCount((int) projection.getMemberCount())
-                         .expenseCount((int) projection.getExpenseCount())
-                         .categoryCount((int) projection.getCategoryCount())
+                         .memberCount(Math.toIntExact(projection.getMemberCount()))
+                         .expenseCount(Math.toIntExact(projection.getExpenseCount()))
+                         .categoryCount(Math.toIntExact(projection.getCategoryCount()))
                          .build();
   }
 }
